Separate interface from implementation as per the normal convention
