+++ CMakeLists.txt
@@ -47,6 +47,9 @@ endif()
 # Specify the C++ standard
 set(CMAKE_CXX_STANDARD 14)
 
+include_directories(${CMAKE_CURRENT_SOURCE_DIR}/include)
+
+add_subdirectory(include)
 add_subdirectory(src)
 
 enable_testing()
