+++ include/geometry/CMakeLists.txt
@@ -1,0 +1,1 @@
+set(GEOMETRY_HDR_FILES Size.h Vector2.h BoundingBox.h)
