+++ unit_tests/geometry/CMakeLists.txt
@@ -1,5 +1,3 @@
-include_directories(../../src/Geometry)
-
 set(TEST_GEOMETRY_SRC_FILES BoundingBox.cpp)
 
 source_group(test\\geometry FILES ${TEST_GEOMETRY_SRC_FILES})
